refactor: rename errors.ts to error.ts in validator and serializer
