+++ src/index.ts
@@ -7,7 +7,7 @@ export {
   type SchemaTransformOptions,
 } from './openapi/schema-transform.js';
 export { jsonSchemaToOAS, type ZodToJsonConfig } from './openapi/zod-to-openapi.js';
-export { ResponseSerializationError } from './serializer/errors.js';
+export { ResponseSerializationError } from './serializer/error.js';
 export {
   createFastSerializerCompiler,
   createParseSerializerCompiler,
@@ -23,5 +23,5 @@ export type {
   FastifyPluginCallbackZod,
   RouteHandler,
 } from './types.js';
-export { RequestValidationError } from './validator/errors.js';
+export { RequestValidationError } from './validator/error.js';
 export { validatorCompiler } from './validator/validator.js';
